+++ src/components/custom/popups/popup-content-display.tsx
